Update to 3.18 kernel
 - Commented out PLAT_PHYS_OFFSET in arch/arm/mach-xm580/include/mach/memory.h
 - Disabled SATA driver for now due to changes in the kernel

+++ drivers/staging/iio/meter/ade7759.c
@@ -124,7 +124,7 @@ static int ade7759_spi_read_reg_40(struct device *dev,
 				reg_address);
 		goto error_ret;
 	}
-	*val = ((u64)st->rx[1] << 32) | ((u64)st->rx[2] << 24) |
+	*val = ((u64)st->rx[1] << 32) | (st->rx[2] << 24) |
 		(st->rx[3] << 16) | (st->rx[4] << 8) | st->rx[5];
 
 error_ret:
@@ -217,6 +217,7 @@ static int ade7759_reset(struct device *dev)
 {
 	int ret;
 	u16 val;
+
 	ade7759_spi_read_reg_16(dev,
 			ADE7759_MODE,
 			&val);
@@ -280,6 +281,7 @@ static int ade7759_set_irq(struct device *dev, bool enable)
 {
 	int ret;
 	u8 irqen;
+
 	ret = ade7759_spi_read_reg_8(dev, ADE7759_IRQEN, &irqen);
 	if (ret)
 		goto error_ret;
@@ -340,6 +342,7 @@ static ssize_t ade7759_read_frequency(struct device *dev,
 	int ret;
 	u16 t;
 	int sps;
+
 	ret = ade7759_spi_read_reg_16(dev,
 			ADE7759_MODE,
 			&t);
